pyproject.toml
# ────────────── build backend ──────────────
[build-system]
requires = [
  "setuptools>=80",      # modern PEP 621 support
#  "wheel",
  "setuptools-scm>=8"    # derive version from Git tags
]
build-backend = "setuptools.build_meta"

# ────────────── project metadata (PEP 621) ──────────────
[project]
name = "pySingleCellNet"
dynamic = ["version"]          
description = "Single-cell classification and analysis, optimized for embryonic and fetal development"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["single cell", "cell typing", "classification", "embryo", "embryoid", "gastruloid"]
authors = [
  {name = "Patrick Cahan", email = "patrick.cahan@gmail.com"},
  {name = "Yuqi Tan",       email = "ytan19@jhmi.edu"}
]

dependencies = [
  "pandas",
  "numpy",
  "scikit-learn",
  "scikit-misc",
  "scanpy",
  "statsmodels",
  "scipy",
  "matplotlib",
  "seaborn",
  "umap-learn",
  "mygene",
  "palettable",
  "gseapy",
  "alive_progress",
  "python-igraph",
  "marsilea"
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.urls]
Documentation = "https://cahanlab-pysinglecellnet.readthedocs-hosted.com/"
Source        = "https://github.com/CahanLab/PySingleCellNet"
Repository    = "https://github.com/CahanLab/PySingleCellNet.git"
Issues        = "https://github.com/CahanLab/PySingleCellNet/issues"
Changelog     = "https://github.com/CahanLab/PySingleCellNet/tree/master/docs/CHANGELOG.md"


# ────────────── setuptools-specific settings ──────────────
#[tool.setuptools]
#package-dir = {"" = "src"}              
#packages = ["pySingleCellNet"]

# Alternative: use the finder (uncomment if you prefer)
[tool.setuptools.packages.find]
where = ["src"]
exclude = ["docs"]

# ────────────── setuptools-scm tweaks (optional) ──────────────
[tool.setuptools_scm]
version_scheme = "only-version"
local_scheme   = "no-local-version"
write_to       = "src/pySingleCellNet/_version.py"

# ────────────── pytest ──────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
